+++ resources/js/components/index.ts
@@ -11,3 +11,4 @@ export { default as ToolCard } from './ToolCard.vue';
 export { default as SiteGuideModal } from './SiteGuideModal.vue';
 export { default as FoldedCornerLink } from './FoldedCornerLink.vue';
 export { default as RotatingText } from './RotatingText.vue';
+export { default as ToolsGuideModal } from './ToolsGuideModal.vue';